+++ js/src/components/google-combo-account-card/connect-ads/connect-ads-footer.js
@@ -9,14 +9,12 @@ import { __ } from '@wordpress/i18n';
 import AppButton from '.~/components/app-button';
 
 /**
+ * Footer component.
  *
- * @param {Object} props Props.
- * @param {boolean} props.isLoading Whether the footer is in a loading state.
- * @param {Function} props.onCreateNew Callback to handle the new ads account creation.
  * @return {JSX.Element} Footer component.
  */
-const ConnectAdsFooter = ( { isLoading, onCreateNew = () => {} } ) => (
-	<AppButton isTertiary disabled={ isLoading } onClick={ onCreateNew }>
+const ConnectAdsFooter = () => (
+	<AppButton isTertiary>
 		{ __(
 			'Or, create a new Google Ads account',
 			'google-listings-and-ads'
