+++ js/src/components/google-combo-account-card/connect-ads/index.js
@@ -2,4 +2,4 @@
  * Internal dependencies
  */
 
-export { default as ConnectAds } from './connect-ads';
+export { default } from './connect-ads';
